requirements.txt
Django==4.2.7
openai==1.3.5
anthropic==0.7.8
python-dotenv==1.0.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
